fix: Display env variables on Nav

+++ src/lib/components/FileList.svelte
@@ -37,7 +37,9 @@
 				<th class="text-center">Name</th>
 				<th class="text-center">Created Time</th>
 				<th class="text-center">Modified Time</th>
-				<th>Delete</th>
+				{#if import.meta.env.VITE_WRITE_PROTECTED === 'false'}
+					<th>Delete</th>
+				{/if}
 			</tr>
 		</thead>
 		<tbody>
@@ -85,28 +87,30 @@
 					{/if}
 					<td class="text-center">{createdTime}</td>
 					<td class="text-center">{modifiedTime}</td>
-					<td>
-						<a
-							href="{sub}{file}"
-							on:click|preventDefault={(e) => handleDelete(e.target)}
-							target="_self"
-						>
-							<svg
-								xmlns="http://www.w3.org/2000/svg"
-								class="h-6 w-6"
-								fill="none"
-								viewBox="0 0 24 24"
-								stroke="currentColor"
+					{#if import.meta.env.VITE_WRITE_PROTECTED === 'false'}
+						<td>
+							<a
+								href="{sub}{file}"
+								on:click|preventDefault={(e) => handleDelete(e.target)}
+								target="_self"
 							>
-								<path
-									stroke-linecap="round"
-									stroke-linejoin="round"
-									stroke-width="2"
-									d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16"
-								/>
-							</svg>
-						</a>
-					</td>
+								<svg
+									xmlns="http://www.w3.org/2000/svg"
+									class="h-6 w-6"
+									fill="none"
+									viewBox="0 0 24 24"
+									stroke="currentColor"
+								>
+									<path
+										stroke-linecap="round"
+										stroke-linejoin="round"
+										stroke-width="2"
+										d="M19 7l-.867 12.142A2 2 0 0116.138 21H7.862a2 2 0 01-1.995-1.858L5 7m5 4v6m4-6v6m1-10V4a1 1 0 00-1-1h-4a1 1 0 00-1 1v3M4 7h16"
+									/>
+								</svg>
+							</a>
+						</td>
+					{/if}
 				</tr>
 			{/each}
 		</tbody>
